split up EditJurisdiction
